leetcode/src/main/java/_54SpiralMatrix.java
import java.util.LinkedList;
import java.util.List;

public class _54SpiralMatrix {

    public List<Integer> spiralOrder(int[][] matrix) {

        List<Integer> list = new LinkedList<>();
        if (matrix.length == 0) {
            return list;
        }

        int top = 0;
        int left = 0;
        int right = matrix[0].length - 1;
        int bottom = matrix.length - 1;

        int side = 0;

        do {
            side = side % 4;
            if (side == 0) {
                for (int i = left; i <= right; i++) {
                    list.add(matrix[top][i]);
                }
                top++;
            } else if (side == 1) {
                for (int i = top; i <= bottom; i++) {
                    list.add(matrix[i][right]);
                }
                right--;
            } else if (side == 2) {
                for (int i = right; i >= left; i--) {
                    list.add(matrix[bottom][i]);
                }
                bottom--;
            } else {
                for (int i = bottom; i >= top; i--) {
                    list.add(matrix[i][left]);
                }
                left++;
            }

            side++;
        } while (top <= bottom && left <= right);

        return list;
    }

}
